requirements.txt
Flask==2.1.1
requests==2.32.3
Werkzeug==2.0.3
discord.py==2.4.0
aiohttp==3.10.5
PyPDF2==3.0.1
python-dotenv==1.0.0
python-docx==1.1.2
google-genai==1.62.0
psycopg[binary]==3.3.3
PyNaCl==1.5.0
discord==2.3.2
discord-ext-voice-recv
